chore(quality): extract ankiConnect + Notion block fetcher factories
Remove the 2.1% Sonar-flagged duplication on PR #2042 by hoisting two
factories that were inlined at every call site:

- ankiConnectFactory in src/services/ankify/buildAnkiConnectClient.ts —
  used at four sites (AnkifyRouter, AnkifyWebhookRouter, and twice in
  data_layer/index.ts)
- notionBlockChildrenFetcherFactory in
  src/services/ankify/notionBlockChildrenFetcher.ts — used at three sites
  (AnkifyRouter SyncNotionPage, AnkifyWebhookRouter, data_layer SyncNotionPage)

The Notion page-meta fetcher and the larger notionDatabaseClient lambda
each appear at only two sites — leaving them inlined per the "wait for
three occurrences before extracting" rule in CLAUDE.md.

Use cases keep receiving the same factory shapes, so existing unit tests
still inject their own factories and don't need updates.

Follow-up SonarQube items deferred (require per-case judgment, not bulk
replacement): S3776 (cognitive complexity, 6), S3358 (nested ternaries,
5), S7735 (negated conditions, 15), S2004 (function nesting, 2).
