+++ src/services/ankify/buildAnkiConnectClient.ts
@@ -1,0 +1,13 @@
+import { AnkiConnectClient, buildAnkiConnectUrl } from './AnkiConnectClient';
+
+export const ankiConnectFactory = (
+  host: string,
+  port: number,
+  apiKey: string | null
+): AnkiConnectClient =>
+  new AnkiConnectClient(
+    buildAnkiConnectUrl(host, port),
+    undefined,
+    undefined,
+    apiKey
+  );
